create console_output folder
