+++ javascript/console_output/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "console_output",
+  "version": "1.0.0",
+  "description": "",
+  "main": "console_output.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "seoulrobotics",
+  "license": "ISC",
+  "dependencies": {
+    "sensr_js_sdk": "file:../javascript_sdk",
+  }
+}
